Migrate TodoList to a Fragment
