+++ app/src/main/res/layout/todo_list_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.todolist.TodoListActivity">
+    tools:context=".ui.todolist.TodoListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/todoListRecyclerView"
