+++ app/src/main/res/navigation/app_navigation.xml
@@ -1,6 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/app_navigation.xml">
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/app_navigation.xml"
+    app:startDestination="@id/todoListFragment">
 
+    <fragment
+        android:id="@+id/todoListFragment"
+        android:name="com.zachtib.simpletodo.ui.todolist.TodoListFragment"
+        android:label="todo_list_fragment"
+        tools:layout="@layout/todo_list_fragment" />
 </navigation>
